Simplify grm payment monitoring
