+++ src/main/kotlin/dsx/bps/DBclasses/core/InvoiceTable.kt
@@ -1,15 +1,15 @@
-package dsx.bps.DBclasses
+package dsx.bps.DBclasses.core
 
 import dsx.bps.core.datamodel.Currency
-import dsx.bps.core.datamodel.PaymentStatus
+import dsx.bps.core.datamodel.InvoiceStatus
 import org.jetbrains.exposed.dao.IntIdTable
 import org.jetbrains.exposed.sql.Column
 import java.math.BigDecimal
 
-object PaymentTable: IntIdTable() {
-    val status = enumeration("status", PaymentStatus::class)
-    val fee: Column<BigDecimal?> = decimal("fee", 30, 15).nullable()
-    val paymentId: Column<String> = varchar("paymentId", 100)
+object InvoiceTable: IntIdTable() {
+    val status = enumeration("status", InvoiceStatus::class)
+    val received: Column<BigDecimal> = decimal("received", 30, 15)
+    val invoiceId: Column<String> = varchar("invoiceId", 100)
     val currency = enumeration("currency", Currency::class)
     val amount: Column<BigDecimal> = decimal("amount", 30, 15)
     val address: Column<String> = varchar("address", 100)
