+++ src/main/kotlin/dsx/bps/DBclasses/crypto/xrp/XrpTxEntity.kt
@@ -1,5 +1,6 @@
-package dsx.bps.DBclasses
+package dsx.bps.DBclasses.crypto.xrp
 
+import dsx.bps.DBclasses.core.TxEntity
 import org.jetbrains.exposed.dao.EntityID
 import org.jetbrains.exposed.dao.IntEntity
 import org.jetbrains.exposed.dao.IntEntityClass
