+++ src/main/kotlin/dsx/bps/core/PaymentProcessor.kt
@@ -62,7 +62,7 @@ class PaymentProcessor(
             processing
                 .mapNotNull { id -> payments[id] }
                 .forEach { pay ->
-                    val tx = manager.updatePaymentTxStatus(pay)
+                    val tx = manager.getTx(pay.currency, pay.txid)
                     if (match(pay, tx)) {
                         when (tx.status()) {
                             TxStatus.VALIDATING -> {
